requirements.txt
flask
python-dotenv
psycopg2-binary

langchain
langchain-core
langchain-huggingface
huggingface-hub